refactor(spells): extract shared spell description utilities
Extract duplicate spell description fetching and enrichment logic from
SpellCard.svelte.ts and SpellReferenceCard.svelte.ts into a shared
utility module (spellDescription.ts).
