+++ src/view/dialogs/components/characterCreator/SpellCard.svelte.ts
@@ -1,37 +1,8 @@
 import type { SpellIndexEntry } from '#utils/getSpells.js';
-
-/**
- * Enriches a text segment using Foundry's TextEditor
- */
-async function enrichText(text: string): Promise<string> {
-	return foundry.applications.ux.TextEditor.implementation.enrichHTML(text);
-}
-
-/**
- * Fetches the full spell document and extracts the description
- */
-async function fetchSpellDescription(uuid: string): Promise<string> {
-	try {
-		const spell = (await fromUuid(uuid)) as Item | null;
-		if (!spell) return '';
-
-		const system = spell.system as {
-			description?: { baseEffect?: string; higherLevelEffect?: string; upcastEffect?: string };
-		};
-		const desc = system?.description;
-
-		// Combine description parts
-		const parts: string[] = [];
-		if (desc?.baseEffect) parts.push(desc.baseEffect);
-		if (desc?.higherLevelEffect)
-			parts.push(`<p><strong>Higher Levels:</strong> ${desc.higherLevelEffect}</p>`);
-		if (desc?.upcastEffect) parts.push(`<p><strong>Upcast:</strong> ${desc.upcastEffect}</p>`);
-
-		return parts.join('');
-	} catch {
-		return '';
-	}
-}
+import {
+	enrichSpellText,
+	fetchSpellDescriptionByUuid,
+} from '../../../../utils/spellDescription.js';
 
 /**
  * Creates reactive state for the SpellCard component
@@ -50,9 +21,9 @@ export function createSpellCardState(getSpell: () => SpellIndexEntry) {
 
 		isLoading = true;
 		const spell = getSpell();
-		const rawDescription = await fetchSpellDescription(spell.uuid);
+		const rawDescription = await fetchSpellDescriptionByUuid(spell.uuid);
 		if (rawDescription) {
-			enrichedDescription = await enrichText(rawDescription);
+			enrichedDescription = await enrichSpellText(rawDescription);
 		}
 		isLoading = false;
 	}
